rename: change shared to base
