+++ app/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
-import { PATH } from "@/shared/constants/path";
+import { PATH } from "@/base/constants/path";
 
 export default function MainScreen() {
   redirect(PATH.dashboard);
