+++ base/components/ui/button.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Slot } from "@radix-ui/react-slot";
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/shared/utils";
+import { cn } from "@/base/utils";
 import { LoadingSpinner } from "./spinner";
 
 const buttonVariants = cva(
