+++ base/components/ui/card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/shared/utils/index"
+import { cn } from "@/base/utils/index"
 
 function Card({ className, ...props }: React.ComponentProps<"div">) {
   return (
