+++ base/components/ui/progress.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 import * as ProgressPrimitive from "@radix-ui/react-progress"
 
-import { cn } from "@/shared/utils/index"
+import { cn } from "@/base/utils/index"
 
 function Progress({
   className,
