+++ entities/interview/apis/apis.ts
@@ -1,4 +1,4 @@
-import { fetchClient } from "@/shared/fetch";
+import { fetchClient } from "@/base/fetch";
 import type { InterviewRecord } from "../models/types";
 
 export const getRecentInterviewRecords = async (): Promise<
