+++ modules/dashboard/ui/dashboard-header.tsx
@@ -1,22 +1,26 @@
 import Link from "next/link";
 import Image from "next/image";
 
-import { Calendar, User } from "lucide-react";
-import { Button, Skeleton } from "@/shared/components/ui";
+import { Calendar, MessageCircle, User } from "lucide-react";
+import { Button, Skeleton } from "@/base/components/ui";
 import { UserFiledBadge } from "@/entities/user/ui";
 
 const DashboardHeader = () => {
   return (
-    <header className="bg-white border-b border-slate-200 px-4 py-3 sticky top-0 z-10">
+    <header className="bg-white/80 backdrop-blur-sm border-b border-slate-200/50 px-4 py-3 sticky top-0 z-10">
       <div className="max-w-6xl mx-auto flex items-center justify-between">
         <div className="flex items-center gap-2">
-          <Image
+          {/* TODO: change logo */}
+          {/* <Image
             src="/images/talkterview-main-logo.png"
             alt="logo"
             width={100}
             height={100}
             className="w-12 h-12"
-          />
+          /> */}
+          <div className="w-10 h-10 bg-gradient-to-r from-blue-600 to-purple-600 rounded-lg flex items-center justify-center">
+            <MessageCircle className="w-6 h-6 text-white" />
+          </div>
           <h1 className="text-xl font-bold text-gray800">톡터뷰</h1>
 
           <UserFiledBadge />
